+++ Lib/fontbakery/checks/conditions.py
