+++ Lib/fontbakery/checks/vendorspecific/adobefonts.py
