+++ Lib/fontbakery/checks/vendorspecific/googlefonts/copyright.py
