+++ Lib/fontbakery/checks/vendorspecific/googlefonts/hinting.py
