+++ Lib/fontbakery/checks/vendorspecific/microsoft/__init__.py
